refactor(media): split transcription draft support
